Refactor the attendants actions on GroupControllers into separate controllers to keep the Group controller clean.
