+++ app/helpers/attendants_helper.rb
@@ -1,0 +1,2 @@
+module AttendantsHelper
+end
